Major structural changes and major code refactoring. Code not currently functional.

+++ Mhotivo/Controllers/AcademicYearController.cs
@@ -85,7 +85,7 @@ namespace Mhotivo.Controllers
 
         [HttpGet]
         [AuthorizeAdmin]
-        public ActionResult Edit(int id)
+        public ActionResult Edit(long id)
         {
             var academicYear = _academicYearRepository.GetById(id);
             var academicYearModel = new AcademicYearEditModel
